tests/factories/category_factory.py
import factory
from faker import Faker
from factory.alchemy import SQLAlchemyModelFactory

from src.adapters.driven.repositories.models.category_model import CategoryModel

fake = Faker()

class CategoryFactory(SQLAlchemyModelFactory):

    class Meta:
        model = CategoryModel
        sqlalchemy_session_persistence = "commit"

    name = factory.LazyAttribute(lambda _: fake.unique.word().capitalize())
    description = factory.LazyAttribute(lambda _: fake.sentence(nb_words=10))
